fix: various fixes due to async panicks
- Fix graceful shutdown by changing the way `dropper_wrapper_fn` works
to use cancel tokens rather than `block_on`.
- removed several `Drop` implementations that previously used `block_on`
and replaced them with handling  `dropper_wrapper_fn`
- Fixed get realm list from DB error due to `population` field being f64
when it should be a positive f32.
- Fixed `WrongPass` config types over `BanType`, `Logging`.
- Implemented Log Appender truncate on open and backup before open
Use those as defaults.

+++ crates/azothacore-server/src/shared/realms.rs
@@ -68,7 +68,7 @@ pub struct Realm {
     pub flag:                   FlagSet<RealmFlags>,
     pub timezone:               u8,
     pub allowed_security_level: AccountTypes,
-    pub population_level:       f64,
+    pub population_level:       f32,
 }
 
 impl Realm {
